help.dbh.wtf instead of desc
I did not add this link as the hyperlink to increase changes of anyone to see that. In comparison with the old text docs give a way better explanation and let people know that it exist(since it has more than just commands). You may suggest what to change on the docs page. Maybe soon add keyword detect for the bot so there will be less people who ask for help?